refactor: consolidate type definitions for better organization
- Moved `Invitation` type from `src/types/invitation.ts` to `src/hooks/types.ts`.
- Moved transaction-related types from `src/types/transaction.ts` to `src/sections/finance/types.ts`.
- Updated imports to use the new type locations across files: `notification-item.tsx`, `finance-transactions-history.tsx`, `use-notifications.ts`, `use-referrals.ts`, and `notifications/index.ts`.
- Adjusted `SubscribeToUnlockCardProps` to use a more specific type from `lens-protocol`.

This change improves type management by consolidating related definitions, reducing redundancy, and enhancing maintainability.
