+++ src/hooks/use-referrals.ts
@@ -1,9 +1,6 @@
 import { useState } from 'react';
 import { useSelector } from 'react-redux';
-import emailjs from '@emailjs/browser';
-
-import { GLOBAL_CONSTANTS } from '@src/config-global';
-import { Invitation } from '@src/types/invitation'; // <-- Importing from separate types file
+import { Invitation } from '@src/hooks/types';
 import {
   fetchInvitations as fetchInvitationsAction,
   checkIfMyEmailHasPendingInvite as checkIfMyEmailHasPendingInviteAction,
@@ -13,6 +10,8 @@ import {
   sendInvitation as sendInvitationAction,
   acceptOrCreateInvitationForUser as acceptOrCreateInvitationForUserAction,
 } from '@src/utils/supabase-actions';
+import { GLOBAL_CONSTANTS } from '@src/config-global';
+import emailjs from '@emailjs/browser';
 
 /**
  * The type for sending emails through EmailJS.
