+++ src/sections/finance/types.ts
@@ -21,3 +21,14 @@ export interface RightSidebarProps {
   following: Profile[] | null | undefined;
   loadingProfiles: boolean;
 }
+
+export type IOrderTableFilterValue = string | Date | null;
+
+export interface IOrderTableFilters {
+  status: string;
+}
+
+export const TRANSACTIONS_TYPES = [
+  { value: 'transferTo', label: 'Income' },
+  { value: 'transferFrom', label: 'Outcomes' },
+];
